+++ stratum/hal/lib/tdi/dpdk/BUILD
@@ -22,22 +22,38 @@ package(
 )
 
 target_defines = [
-    "DPDK_TARGET",      # shouldn't be necessary
+    "DPDK_TARGET",  # shouldn't be necessary
 ]
+
+stratum_cc_library(
+    name = "dpdk_port_manager",
+    srcs = ["dpdk_port_manager.cc"],
+    hdrs = ["dpdk_port_manager.h"],
+    deps = [
+        "//stratum/glue:integral_types",
+        "//stratum/glue:logging",
+        "//stratum/glue/status",
+        "//stratum/glue/status:statusor",
+        "//stratum/hal/lib/common:common_cc_proto",
+        "//stratum/hal/lib/tdi:tdi_sde_headers",
+        "//stratum/lib/channel",
+        "@com_google_absl//absl/synchronization",
+        "@com_google_absl//absl/time",
+    ],
+)
 
 stratum_cc_library(
     name = "dpdk_chassis_manager",
     srcs = ["dpdk_chassis_manager.cc"],
     hdrs = ["dpdk_chassis_manager.h"],
-    defines = target_defines,
     deps = [
-	":dpdk_port_config",
-	":dpdk_port_constants",
+        ":dpdk_port_config",
+        ":dpdk_port_constants",
+        "//stratum/glue:integral_types",
+        "//stratum/glue:logging",
         "//stratum/glue/gtl:map_util",
         "//stratum/glue/status",
         "//stratum/glue/status:statusor",
-        "//stratum/glue:integral_types",
-        "//stratum/glue:logging",
         "//stratum/hal/lib/common:common_cc_proto",
         "//stratum/hal/lib/common:constants",
         "//stratum/hal/lib/common:switch_interface",
@@ -60,12 +76,16 @@ stratum_cc_library(
 stratum_cc_library(
     name = "dpdk_port_config",
     srcs = ["dpdk_port_config.cc"],
-    hdrs = ["dpdk_port_config.h"],
+    hdrs = [
+        "dpdk_port_config.h",
+        "dpdk_port_manager.h",
+    ],
     deps = [
-	":dpdk_port_constants",
-        "//stratum/glue/status",
+        ":dpdk_port_constants",
         "//stratum/glue:integral_types",
+        "//stratum/glue/status",
         "//stratum/hal/lib/common:common_cc_proto",
+        "//stratum/hal/lib/tdi:macros",
         "//stratum/hal/lib/tdi:tdi_sde_interface",
         "//stratum/public/lib:error",
         "@com_github_google_glog//:glog",
@@ -111,16 +131,16 @@ stratum_cc_library(
     defines = target_defines,
     deps = [
         ":dpdk_chassis_manager",
-        "//stratum/glue/gtl:map_util",
-        "//stratum/glue/status:status_macros",
         "//stratum/glue:integral_types",
         "//stratum/glue:logging",
+        "//stratum/glue/gtl:map_util",
+        "//stratum/glue/status:status_macros",
+        "//stratum/hal/lib/common:switch_interface",
         "//stratum/hal/lib/tdi:tdi_cc_proto",
         "//stratum/hal/lib/tdi:tdi_node",
         "//stratum/hal/lib/tdi:tdi_pipeline_utils",
         "//stratum/hal/lib/tdi:tdi_sde_interface",
         "//stratum/hal/lib/tdi:utils",
-        "//stratum/hal/lib/common:switch_interface",
         "//stratum/lib:constants",
         "//stratum/lib:macros",
         "@com_google_absl//absl/memory",
@@ -136,11 +156,11 @@ stratum_cc_test(
     deps = [
         ":dpdk_chassis_manager",
         ":test_main",
+        "//stratum/glue:integral_types",
         "//stratum/glue/gtl:map_util",
         "//stratum/glue/status",
         "//stratum/glue/status:status_test_util",
         "//stratum/glue/status:statusor",
-        "//stratum/glue:integral_types",
         "//stratum/hal/lib/common:common_cc_proto",
         "//stratum/hal/lib/common:constants",
         "//stratum/hal/lib/common:utils",
@@ -160,8 +180,8 @@ stratum_cc_library(
     name = "dpdk_chassis_manager_mock",
     testonly = 1,
     hdrs = ["dpdk_chassis_manager_mock.h"],
-    defines = target_defines,
     arches = HOST_ARCHES,
+    defines = target_defines,
     deps = [
         ":dpdk_chassis_manager",
         "@com_google_googletest//:gtest",
@@ -193,18 +213,18 @@ stratum_cc_library(
         "//stratum/hal/lib/common:error_buffer",
         "//stratum/hal/lib/common:p4_service",
         "//stratum/hal/lib/common:switch_interface",
-        "//stratum/lib/security:auth_policy_checker",
-	"//stratum/lib/security:credentials_manager",
         "//stratum/lib:constants",
         "//stratum/lib:macros",
         "//stratum/lib:utils",
+        "//stratum/lib/security:auth_policy_checker",
+        "//stratum/lib/security:credentials_manager",
         "@com_github_google_glog//:glog",
+        "@com_github_grpc_grpc//:grpc++",
         "@com_google_absl//absl/base:core_headers",
         "@com_google_absl//absl/container:flat_hash_map",
         "@com_google_absl//absl/memory",
         "@com_google_absl//absl/strings",
         "@com_google_absl//absl/synchronization",
-        "@com_github_grpc_grpc//:grpc++",
     ],
 )
 
@@ -220,9 +240,9 @@ stratum_cc_test(
         "//stratum/glue/net_util:ports",
         "//stratum/glue/status:status_test_util",
         "//stratum/hal/lib/common:switch_mock",
+        "//stratum/lib:utils",
         "//stratum/lib/security:auth_policy_checker_mock",
         "//stratum/lib/test_utils:matchers",
-        "//stratum/lib:utils",
         "//stratum/public/lib:error",
         "@com_github_google_glog//:glog",
         "@com_google_absl//absl/strings",
@@ -236,9 +256,9 @@ stratum_cc_test(
         "dpdk_port_config_test.cc",
     ],
     deps = [
-	":dpdk_port_config",
-	"//stratum/hal/lib/tdi:tdi_sde_interface",
+        ":dpdk_port_config",
         ":test_main",
+        "//stratum/hal/lib/tdi:tdi_sde_interface",
         "@com_google_googletest//:gtest",
     ],
 )
@@ -259,8 +279,8 @@ stratum_cc_test(
         "//stratum/hal/lib/p4:p4_table_mapper_mock",
         "//stratum/hal/lib/tdi:tdi_node_mock",
         "//stratum/hal/lib/tdi:tdi_sde_mock",
-        "//stratum/lib/channel:channel_mock",
         "//stratum/lib:utils",
+        "//stratum/lib/channel:channel_mock",
         "//stratum/lib/test_utils:matchers",
         "@com_google_absl//absl/memory",
         "@com_google_googletest//:gtest",
